move utility function into another package, make more robust
